requirements.txt
requests
python-whois
dnspython
networkx
matplotlib